Refactor the inmemory server to allow for some better integration testing
