+++ eventbus-inmemory/tests/simple-exchange.rs
@@ -3,20 +3,31 @@
  * the eventbus and a websocket client
  */
 
+extern crate eventbus_inmemory;
+
 use async_std::task;
-
 use log::*;
-use otto_eventbus::server::*;
-
 use tungstenite::client::AutoStream;
 use tungstenite::handshake::client::Response;
 use tungstenite::*;
 use url::Url;
 
+fn ws_connect() -> Result<(WebSocket<AutoStream>, Response)> {
+    return connect(Url::parse("ws://127.0.0.1:8105/").unwrap());
+}
 
 #[async_std::test]
 async fn simple_connect() -> std::io::Result<()> {
+    pretty_env_logger::init();
+    std::thread::spawn(|| {
+        println!("YOLO");
+        smol::run(
+            eventbus_inmemory::run_server("127.0.0.1:8105".to_string())
+        )
+    });
 
-    assert!(false);
+    let (mut socket, _response) = ws_connect().unwrap();
+    println!("Connected to the server");
+
     Ok(())
 }
